Added routes.js
